typings install ejs --source=dt
