+++ src/AzureCloudService/WebRole1/typings/index.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="modules/ejs/index.d.ts" />
