+++ src/AzureCloudService/WebRole1/typings/modules/ejs/typings.json
@@ -1,0 +1,8 @@
+{
+  "resolution": "main",
+  "tree": {
+    "src": "https://raw.githubusercontent.com/DefinitelyTyped/DefinitelyTyped/cd4debea25a280da0808d4ff2ca5a4bdb34bd28b/ejs/index.d.ts",
+    "raw": "registry:dt/ejs#2.3.3+20161128184045",
+    "typings": "https://raw.githubusercontent.com/DefinitelyTyped/DefinitelyTyped/cd4debea25a280da0808d4ff2ca5a4bdb34bd28b/ejs/index.d.ts"
+  }
+}
